WIP tests...having issues with responses from servers...
WHOIS servers seem to not like getting hammered with requests while testing.
That in mind, it seems like I may need to come up with a concept to mock the server.
That will be tricky AF since I need to mock a TLD essentially - and make the mock server give different results per requesting "Server".